Finish2

+++ src/main/java/pro/sky/telegrambot/listener/TelegramBotUpdatesListener.java
@@ -4,6 +4,7 @@ import com.pengrad.telegrambot.TelegramBot;
 import com.pengrad.telegrambot.UpdatesListener;
 import com.pengrad.telegrambot.model.Update;
 import com.pengrad.telegrambot.request.SendMessage;
+import com.pengrad.telegrambot.response.SendResponse;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.scheduling.annotation.Scheduled;
@@ -23,7 +24,7 @@ public class TelegramBotUpdatesListener implements UpdatesListener {
 
     public static final DateTimeFormatter DATE_TIME_FORMATTER = DateTimeFormatter.ofPattern("dd.MM.yyyy HH:mm");
     private final Logger logger = LoggerFactory.getLogger(TelegramBotUpdatesListener.class);
-    private static final String REGEX = "([0-9.:\\s]{16})(\\s)([\\W+]+)";
+    private static final String REGEX = "([0-9\\.\\:\\s]{16})(\\s)([\\W+]+)";
     private final Pattern pattern = Pattern.compile(REGEX);
 
     private final TelegramBot telegramBot;
@@ -43,21 +44,30 @@ public class TelegramBotUpdatesListener implements UpdatesListener {
     @Override
     public int process(List<Update> updates) {
         updates.forEach(update -> {
-            logger.info("Processing update: {}", update);
-            if (update.message() == null) {
-                return;
-            }
-            String text = update.message().text();
-            Long chatId = update.message().chat().id();
-            Matcher matcher = pattern.matcher(text);
-            if (matcher.matches()) {
-                NotificationTask task = new NotificationTask();
-                task.setMessage(matcher.group(3));
-                task.setChatId(chatId);
-                task.setExecDate(LocalDateTime.parse(matcher.group(1), DATE_TIME_FORMATTER));
-                notificationTaskRepository.save(task);
-            } else if (text.equals("/start")) {
-                telegramBot.execute(new SendMessage(chatId, "Hello, World!"));
+            try {
+                logger.info("Processing update: {}", update);
+                if (update.message() == null) {
+                    return;
+                }
+                String text = update.message().text();
+                Long chatId = update.message().chat().id();
+                Matcher matcher = pattern.matcher(text);
+                if (matcher.matches()) {
+                    LocalDateTime execDate = LocalDateTime.parse(matcher.group(1), DATE_TIME_FORMATTER);
+                    if (execDate.isBefore(LocalDateTime.now())) {
+                        telegramBot.execute(new SendMessage(chatId, "Укажите дату в будущем"));
+                    } else {
+                        NotificationTask task = new NotificationTask();
+                        task.setMessage(matcher.group(3));
+                        task.setChatId(chatId);
+                        task.setExecDate(LocalDateTime.parse(matcher.group(1), DATE_TIME_FORMATTER));
+                        notificationTaskRepository.save(task);
+                    }
+                } else if (text.equals("/start")) {
+                    telegramBot.execute(new SendMessage(chatId, "Hello, World!"));
+                }
+            } catch (Exception e) {
+                logger.error("Failed to process update {}", update, e);
             }
         });
         return UpdatesListener.CONFIRMED_UPDATES_ALL;
@@ -67,12 +77,17 @@ public class TelegramBotUpdatesListener implements UpdatesListener {
     public void timer() {
         notificationTaskRepository.findAllByExecDateLessThan(LocalDateTime.now()).forEach(
                 task -> {
-                    telegramBot.execute(new SendMessage(task.getChatId(), task.getMessage()));
-                    notificationTaskRepository.delete(task);
+                    SendResponse execute = telegramBot.execute(new SendMessage(task.getChatId(), task.getMessage()));
+                    if (execute.isOk()) {
+                        notificationTaskRepository.delete(task);
+                    } else {
+                        logger.error("Failed to send message to " + task.getChatId());
+                    }
                 }
         );
 
 
     }
-
 }
+
+
